review fix and there is a discussion points
