+++ saltapi/main.py
@@ -13,6 +13,7 @@ from saltapi.web.api.institutions import router as institution_router
 from saltapi.web.api.instruments import router as instruments_router
 from saltapi.web.api.proposal_progress import router as progress_router
 from saltapi.web.api.proposals import router as proposals_router
+from saltapi.web.api.salt_astronomers import router as salt_astronomers_router
 from saltapi.web.api.submissions import router as submissions_router
 from saltapi.web.api.user import router as user_router
 from saltapi.web.api.users import router as users_router
@@ -48,5 +49,6 @@ app.include_router(user_router)
 app.include_router(users_router)
 app.include_router(instruments_router)
 app.include_router(institution_router)
+app.include_router(salt_astronomers_router)
 app.include_router(submissions_router)
 app.include_router(finder_charts_router)
