+++ saltapi/web/api/users.py
@@ -100,17 +100,6 @@ def get_users(
         return user_service.get_users()
 
 
-@router.get(
-    "/salt-astronomers",
-    summary="Get the SALT astronomers",
-    response_model=List[UserListItem],
-)
-def get_salt_astronomers() -> List[Dict[str, Any]]:
-    with UnitOfWork() as unit_of_work:
-        user_service = services.user_service(unit_of_work.connection)
-        return user_service.get_salt_astronomers()
-
-
 @router.get("/{user_id}", summary="Get user details", response_model=User)
 def get_user(
     user_id: int = Path(
